logging: move shard truncation into persist_sink
This commit moves the async persist shard truncation step, performed
when creating a logging dataflow sink, into the async persist_sink
operator. This removes the need for calling tokio's `block_on`, which
previously could lead to deadlocks because the truncate operation needs
to `await` while holding a lock on the `persist_clients` mutex and
timely wouldn't know how to put that tokio task to sleep and schedule
other operators.
